Move IABService back into src/index
